Update main for minimun changes and added TODO

+++ main.py
@@ -3,6 +3,7 @@ from models.models import WeatherData
 from database.schemas import WeatherRequest, WeatherData as WeatherDataSchema
 from services.services import fetch_and_store_weather_data
 from database.context_manager import SessionLocal
+from database.context_manager import database
 from services.services import CITIES
 import json
 
@@ -10,6 +11,9 @@ import json
 #Base.metadata.create_all(bind=engine)
 
 app = FastAPI()
+
+
+#TODO response models
 
 
 @app.post("/weather/")
@@ -20,7 +24,7 @@ async def post_weather(request: WeatherRequest):
         raise HTTPException(status_code=400, detail="User ID already exists")
 
     await fetch_and_store_weather_data(request.user_id)
-    return  {"message": "Weather data collection started"}
+    return {"message": "Weather data collection started"}
 
 
 @app.get("/weather/{user_id}")
@@ -33,12 +37,11 @@ async def get_weather(user_id: str):
     if not record:
         raise HTTPException(status_code=404, detail="User ID not found")
 
-
     # Count the number of elements in the list
     entries_count = len(json.loads(record.data))
 
     # Calculate the percentage of cities uploaded
-    percentage_uploaded = (entries_count / total_cities) * 100
+    percentage_uploaded = int((entries_count / total_cities) * 100)
 
     # Fetch the first entry to maintain the existing functionality
     result = db.query(WeatherData).filter(WeatherData.user_id == user_id).first()
